+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.1.12] - 2021-09-20
+
+### Added
+
+- Added geometry.Spacing
+
+### Fixed
+
+- Fixed calculation of virtual size in scroll views
+
 ## [0.1.11] - 2021-09-12
 
 ### Changed